Fix NoClassDefFoundError and get Magma to run
